upgrade site to gatsby v2 beta
