+++ .gitignore
@@ -1,3 +1,8 @@
-node_modules
-public
+# Project dependencies
 .cache
+node_modules
+yarn-error.log
+
+# Build directory
+/public
+.DS_Store
