+++ README.md
@@ -1,7 +1,26 @@
-# Gatsby Manor
-Here is the source code of the main site. Built with Gatsby of course!
+# gatsby-starter-default
+The default Gatsby starter.
 
-## How to contribute a Starter
-Add your starter information `./src/data/starters.js`.
-Add your starter iamges `./src/images`.
-Make a pull request.
+For an overview of the project structure please refer to the [Gatsby documentation - Building with Components](https://www.gatsbyjs.org/docs/building-with-components/).
+
+## Install
+
+Make sure that you have the Gatsby CLI program installed:
+```sh
+npm install --global gatsby-cli
+```
+
+And run from your CLI:
+```sh
+gatsby new gatsby-example-site
+```
+
+Then you can run it by:
+```sh
+cd gatsby-example-site
+gatsby develop
+```
+
+## Deploy
+
+[![Deploy to Netlify](https://www.netlify.com/img/deploy/button.svg)](https://app.netlify.com/start/deploy?repository=https://github.com/gatsbyjs/gatsby-starter-default)
